Fixed compiling error

+++ TD1/exercice_1/fibonacci.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
-#include <fibonacci.h>
+#include "fibonacci.h"
 
-int fibonacci(int n, int DPFib[]) {
+long long int fibonacci(int n, long long int DPFib[]) {
     if (n <= 1) return n;
 
     if (DPFib[n] != -1) return DPFib[n];
@@ -11,7 +11,7 @@ int fibonacci(int n, int DPFib[]) {
 
 }
 
-int fibonacci_naive(int n) {
+long long int fibonacci_naive(int n) {
     if (n <= 1)
         return n;
     return fibonacci_naive(n - 1) + fibonacci_naive(n - 2);
