+++ TD1/exercice_1/fibonacci.h
@@ -1,2 +1,2 @@
-int fibonacci(int n, int DPFib[]);
-int fibonacci_naive(int n);
+long long int fibonacci(int n, long long int DPFib[]);
+long long int fibonacci_naive(int n);
